+++ src/lib/mema.c
@@ -38,8 +38,8 @@
 #define MAX_SPACES 124
 
 struct memaspace {
-	s32 addr;
-	u32 size;
+	uintptr_t addr;
+	u64 size;
 };
 
 /**
@@ -55,8 +55,8 @@ struct memaheap {
 	struct memaspace end2;
 };
 
-s32 g_MemaHeapStart;
-s32 g_MemaHeapSize;
+uintptr_t g_MemaHeapStart;
+u64 g_MemaHeapSize;
 struct memaheap g_MemaHeap;
 
 #if VERSION == VERSION_PAL_BETA
@@ -65,8 +65,8 @@ u32 g_MemaLeastEverFree = 1000000;
 
 void memaSwap(struct memaspace *a, struct memaspace *b)
 {
-	u32 tempaddr = a->addr;
-	u32 tempsize = a->size;
+	uintptr_t tempaddr = a->addr;
+	u64 tempsize = a->size;
 	a->addr = b->addr;
 	a->size = b->size;
 	b->addr = tempaddr;
@@ -86,7 +86,7 @@ bool memaDefragPass(struct memaheap *heap)
 	struct memaspace *prev = &heap->start;
 	struct memaspace *curr = &heap->spaces[0];
 	struct memaspace *last = &heap->spaces[MAX_SPACES - 1];
-	u32 addr = 0;
+	uintptr_t addr = 0;
 
 	while (curr <= last) {
 		if (curr->size != 0) {
@@ -177,7 +177,7 @@ struct memaspace *memaMakeSlot(struct memaheap *heap)
 	return best;
 }
 
-void _memaFree(s32 addr, s32 size)
+void _memaFree(uintptr_t addr, u64 size)
 {
 	// Choose an index in the spaces array which we'll mark a space as free,
 	// based on how far into the heap the allocation is. This is a rough
@@ -215,7 +215,7 @@ void memaInit(void)
 	// empty
 }
 
-void memaReset(void *heapaddr, u32 heapsize)
+void memaReset(void *heapaddr, u64 heapsize)
 {
 	struct memaspace *space;
 
@@ -398,14 +398,14 @@ void memaPrint(void)
 #endif
 }
 
-void *memaAlloc(u32 size)
+void *memaAlloc(u64 size)
 {
-	u32 addr;
-	u32 diff;
+	uintptr_t addr;
+	uintptr_t diff;
 	s32 i;
 
 	struct memaspace *curr;
-	u32 bestdiff;
+	uintptr_t bestdiff;
 	struct memaspace *best;
 
 	if (1);
@@ -487,7 +487,7 @@ void *memaAlloc(u32 size)
 /**
  * Grow the allocation which currently *ends at* the given address.
  */
-s32 memaGrow(s32 addr, u32 amount)
+uintptr_t memaGrow(uintptr_t addr, u64 amount)
 {
 	struct memaspace *curr = &g_MemaHeap.spaces[0];
 
@@ -512,7 +512,7 @@ found:
 	return addr;
 }
 
-void memaFree(void *addr, s32 size)
+void memaFree(void *addr, u64 size)
 {
 	_memaFree((uintptr_t) addr, size);
 }
@@ -526,7 +526,7 @@ void mema00012cd4(void)
  * Find and return the largest amount of contiguous free space in the pool.
  * ie. the biggest allocation that mema can currently make.
  */
-s32 memaGetLongestFree(void)
+u64 memaGetLongestFree(void)
 {
 	struct memaspace *curr;
 	s32 biggest = 0;
@@ -550,7 +550,7 @@ s32 memaGetLongestFree(void)
 	return 0;
 }
 
-bool memaRealloc(s32 addr, u32 oldsize, u32 newsize)
+bool memaRealloc(uintptr_t addr, u64 oldsize, u64 newsize)
 {
 	if (newsize > oldsize) {
 		if (!memaGrow(addr + oldsize, newsize - oldsize)) {
