+++ src/lib/ultra/io/pidma.c
@@ -1,7 +1,7 @@
 #include <os_internal.h>
 #include "piint.h"
 
-s32 osPiStartDma(OSIoMesg *mb, s32 priority, s32 direction, u32 devAddr, void *dramAddr, u32 size, OSMesgQueue *mq)
+s32 osPiStartDma(OSIoMesg *mb, s32 priority, s32 direction, uintptr_t devAddr, void *dramAddr, u32 size, OSMesgQueue *mq)
 {
 	register s32 ret;
